@@ -361,14 +361,13 @@ public class EditNoteFragment extends Fragment implements CheckableItemAdapter.O
     private void resetBackgroundColors() {
         if (getActivity() != null) {
 
-            MaterialCardView cardView = new MaterialCardView(requireContext());
-            bottomCard.setBackgroundColor(cardView.getCardBackgroundColor().getDefaultColor());
+            bottomCard.setBackgroundColor(ContextCompat.getColor(requireContext(), R.color.note_bottom_background_color_default));
             view.setBackgroundColor(ContextCompat.getColor(requireContext(), R.color.note_background_color_default));
             note_background_color = ContextCompat.getColor(requireContext(), R.color.note_background_color_default);
 
             Window window = requireActivity().getWindow();
             window.addFlags(WindowManager.LayoutParams.FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS);
-            window.setNavigationBarColor(cardView.getCardBackgroundColor().getDefaultColor());
+            window.setNavigationBarColor(ContextCompat.getColor(requireContext(), R.color.note_bottom_background_color_default));
             window.setStatusBarColor(ContextCompat.getColor(requireContext(), R.color.note_background_color_default));
         }
     }
@@ -387,7 +386,7 @@ public class EditNoteFragment extends Fragment implements CheckableItemAdapter.O
         mAdapter.setOnItemClickListener(this);
         String uid = "" + System.currentTimeMillis() + checkableItemList.size();
         checkableItemList.add(new CheckableItem("", false, uid));
-        mAdapter.notifyDataSetChanged();
+        mAdapter.notifyItemInserted(checkableItemList.size() - 1);
 
         addCheckboxBtn.setOnClickListener(v -> {
             String uid1 = "" + System.currentTimeMillis() + checkableItemList.size();
